Replace old sycamore_reactive crate
