+++ packages/sycamore-reactive/Cargo.toml
@@ -10,14 +10,14 @@ readme = "../../README.md"
 repository = "https://github.com/sycamore-rs/sycamore"
 version = "0.9.0-beta.1"
 
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
 [dependencies]
-bumpalo = { version = "3.12.0", features = ["boxed"] }
-indexmap = "1.9.2"
-getrandom = { version = "0.2.8", features = ["js"] }
-hashbrown = "0.13.1"
-serde = { version = "1.0.147", optional = true }
+paste = "1.0.12"
+serde = { version = "1.0.188", optional = true }
 slotmap = "1.0.6"
-smallvec = { version = "1.10.0", features = ["const_generics"] }
+sycamore-reactive-macro = { path = "../sycamore-reactive-macro" }
 
 [features]
 default = []
+nightly = []
