--- packages/sycamore-reactive3/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "sycamore-reactive3"
-categories = ["gui", "wasm", "web-programming"]
-description = "Reactive primitives for Sycamore"
-edition = "2021"
-homepage = "https://github.com/sycamore-rs/sycamore"
-keywords = ["wasm", "gui", "reactive"]
-license = "MIT"
-readme = "../../README.md"
-repository = "https://github.com/sycamore-rs/sycamore"
-version = "0.9.0-beta.1"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-paste = "1.0.12"
-serde = { version = "1.0.188", optional = true }
-slotmap = "1.0.6"
-sycamore-reactive-macro = { path = "../sycamore-reactive-macro" }
-
-[features]
-default = []
-nightly = []
