+++ packages/sycamore-web/Cargo.toml
@@ -17,7 +17,7 @@ indexmap = "1.9.2"
 js-sys = "0.3.60"
 once_cell = "1.16.0"
 sycamore-core = { path = "../sycamore-core", version = "0.9.0-beta.1" }
-sycamore-reactive3 = { path = "../sycamore-reactive3", version = "0.9.0-beta.1" }
+sycamore-reactive = { path = "../sycamore-reactive", version = "0.9.0-beta.1" }
 wasm-bindgen = "0.2.83"
 web-sys = { version = "0.3.60", features = [
 	"console",
