+++ packages/sycamore-web/src/ssr_node.rs
@@ -15,7 +15,7 @@ use sycamore_core::generic_node::{
 };
 use sycamore_core::hydrate::{get_next_id, with_hydration_context};
 use sycamore_core::view::View;
-use sycamore_reactive3::*;
+use sycamore_reactive::*;
 use wasm_bindgen::prelude::*;
 
 use crate::{Html, VOID_ELEMENTS};
